requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
shap>=0.42.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
pillow>=10.0.0
joblib>=1.3.0